app/requirements.txt
pymupdf4llm
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
instructor
groq
requests
pytessaract